requirements.txt
openai>=1.0.0
firecrawl-py>=0.0.1
python-dotenv>=1.0.0

